Fix fast refresh compatibility for calendar
