+++ frontend/src/citizen-frontend/calendar/MonthElem.tsx
@@ -7,12 +7,8 @@ import React from 'react'
 import styled, { css } from 'styled-components'
 
 import type { CitizenCalendarEvent } from 'lib-common/generated/api-types/calendarevent'
-import type {
-  ReservationChild,
-  ReservationResponseDay
-} from 'lib-common/generated/api-types/reservations'
+import type { ReservationResponseDay } from 'lib-common/generated/api-types/reservations'
 import type LocalDate from 'lib-common/local-date'
-import { formatPreferredName } from 'lib-common/names'
 import { Button } from 'lib-components/atoms/buttons/Button'
 import {
   ExpandingInfoBox,
@@ -30,29 +26,6 @@ import type { MonthlyTimeSummary } from './MonthlyHoursSummary'
 import MonthlyHoursSummary from './MonthlyHoursSummary'
 import type { ChildImageData } from './RoundChildImages'
 import { useSummaryInfo } from './hooks'
-
-export function getSummaryForMonth(
-  childData: ReservationChild[],
-  year: number,
-  month: number
-): MonthlyTimeSummary[] {
-  return childData.flatMap(({ monthSummaries, firstName, preferredName }) => {
-    const summaryForMonth = monthSummaries?.find(
-      (monthSummary) =>
-        monthSummary.year === year && monthSummary.month === month
-    )
-    if (!summaryForMonth) {
-      return []
-    }
-    return {
-      name: formatPreferredName({
-        firstName,
-        preferredName
-      }),
-      ...summaryForMonth
-    }
-  })
-}
 
 interface MonthProps {
   calendarMonth: CalendarMonth
@@ -154,27 +127,6 @@ export interface CalendarMonth {
   year: number
   monthNumber: number
   calendarDays: ReservationResponseDay[]
-}
-
-export function groupByMonth(days: ReservationResponseDay[]): CalendarMonth[] {
-  const months: CalendarMonth[] = []
-  let currentMonth: CalendarMonth | undefined = undefined
-  days.forEach((d) => {
-    if (
-      !currentMonth ||
-      currentMonth.year !== d.date.year ||
-      currentMonth.monthNumber !== d.date.month
-    ) {
-      currentMonth = {
-        year: d.date.year,
-        monthNumber: d.date.month,
-        calendarDays: []
-      }
-      months.push(currentMonth)
-    }
-    currentMonth.calendarDays.push(d)
-  })
-  return months
 }
 
 const titleStyles = css`
