+++ frontend/src/citizen-frontend/calendar/RoundChildImages.tsx
@@ -6,12 +6,10 @@ import take from 'lodash/take'
 import React, { useMemo } from 'react'
 import styled, { css } from 'styled-components'
 
-import type { ReservationChild } from 'lib-common/generated/api-types/reservations'
 import type {
   ChildId,
   ChildImageId
 } from 'lib-common/generated/api-types/shared'
-import { formatFirstName } from 'lib-common/names'
 import { fontWeights } from 'lib-components/typography'
 import { theme } from 'lib-customizations/common'
 
@@ -148,14 +146,3 @@ const ChildImageFallback = styled.div<{
   justify-content: center;
   align-items: center;
 `
-
-export const getChildImages = (
-  childData: ReservationChild[]
-): ChildImageData[] =>
-  childData.map((child, index) => ({
-    childId: child.id,
-    imageId: child.imageId,
-    initialLetter: (formatFirstName(child) || '?')[0],
-    colorIndex: index,
-    childName: child.firstName
-  }))
